Update ArbOS syscall support in the aggregator
