+++ packages/arbos-contracts/contracts/ArbSys.sol
@@ -17,6 +17,11 @@ interface ArbSys {
     // or the account sequence number of the given contract
     function getTransactionCount(address account) external view returns (uint256);
 
+    // Return the value of the storage slot for the given account at the given index
+    // This function is only callable from address 0 to prevent contracts from being
+    // able to call it
+    function getStorageAt(address account, uint256 index) external view returns (uint256);
+
     event EthWithdrawal(address indexed destAddr, uint256 amount);
     event ERC20Withdrawal(address indexed destAddr, address indexed tokenAddr, uint256 amount);
     event ERC721Withdrawal(address indexed destAddr, address indexed tokenAddr, uint256 indexed id);
